refactor(Model Traits): restructure model traits
