+++ src/Traits/Bookmark/HasBookmarks.php
@@ -3,53 +3,13 @@
 namespace LaraZeus\Mark\Traits\Bookmark;
 
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\HasMany;
-use LaraZeus\Mark\Facades\Mark;
-use LaraZeus\Mark\NotPassed;
 
 /**
  * @mixin Model
  */
 trait HasBookmarks
 {
-    public function bookmarks(): HasMany
-    {
-        return $this->hasMany(Mark::getBookmarkMorphPivotModel(), 'marker_id');
-    }
-
-    public function hasBookmarked(Model $model): bool
-    {
-        return $this->bookmarks()->whereMorphedTo('markable', $model)->exists();
-    }
-
-    public function markBookmark(Model $markable, bool $value, array | null | NotPassed $metaData = new NotPassed)
-    {
-        $attributes = [
-            'markable_type' => $markable->getMorphClass(),
-            'markable_id' => $markable->getKey(),
-        ];
-
-        $values = [
-            'value' => $value,
-        ];
-
-        if (! $metaData instanceof NotPassed) {
-            $values['metadata'] = $metaData;
-        }
-
-        return $this->bookmarks()->updateOrCreate($attributes, $values);
-    }
-
-    public function unmarkBookmark(Model $markable)
-    {
-        return $this->bookmarks()
-            ->whereMorphedTo('markable', $markable)
-            ->first()
-            ?->delete();
-    }
-
-    public function bookmark(Model $markable, array | null | NotPassed $metaData = new NotPassed)
-    {
-        return $this->markBookmark($markable, true, $metaData);
-    }
+    use Marker\Actions;
+    use Marker\Indicators;
+    use Marker\Relations;
 }
