+++ src/Traits/Bookmark/Marker/Indicators.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Mark\Traits\Bookmark\Marker;
+
+use Illuminate\Database\Eloquent\Model;
+
+trait Indicators
+{
+    public function hasBookmarked(Model $model): bool
+    {
+        return $this->bookmarks()->whereMorphedTo('markable', $model)->exists();
+    }
+}
