+++ src/Traits/Rating/Marker/Actions.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace LaraZeus\Mark\Traits\Rating\Marker;
+
+use Illuminate\Database\Eloquent\Model;
+use LaraZeus\Mark\NotPassed;
+
+trait Actions
+{
+    public function markRating(Model $markable, int $value, array | null | NotPassed $metaData = new NotPassed)
+    {
+        $attributes = [
+            'markable_type' => $markable->getMorphClass(),
+            'markable_id' => $markable->getKey(),
+        ];
+
+        $values = [
+            'value' => $value,
+        ];
+
+        if (! $metaData instanceof NotPassed) {
+            $values['metadata'] = $metaData;
+        }
+
+        return $this->ratings()->updateOrCreate($attributes, $values);
+    }
+
+    public function unmarkRating(Model $markable)
+    {
+        return $this->ratings()
+            ->whereMorphedTo('markable', $markable)
+            ->first()
+            ?->delete();
+    }
+
+    public function rate(Model $markable, int $value, array | null | NotPassed $metaData = new NotPassed)
+    {
+        return $this->markRating($markable, $value, $metaData);
+    }
+}
